+++ Assets/Script/GamePlay/Employee/Data/IEmployee.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public interface IEmployee
 {
-    long ID { get; set; }
+    int ID { get; set; }
     public EmployeeSO _EmployeeType { get; set; }
     string Name { get; set; }
     int Age { get; set; }
